requirements.txt
fastapi==0.101.1
yt-dlp==2023.7.6
ffmpeg-python==0.2.0
uvicorn==0.23.2
